requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.6.1
    # via starlette
asgiref==3.5.2
    # via uvicorn
click==8.1.3
    # via uvicorn
fastapi==0.78.0
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
idna==3.3
    # via anyio
pydantic==1.9.1
    # via
    #   -r requirements.in
    #   fastapi
pyjwt==2.4.0
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
six==1.16.0
    # via python-multipart
sniffio==1.2.0
    # via anyio
starlette==0.19.1
    # via fastapi
typing-extensions==4.2.0
    # via
    #   pydantic
    #   starlette
uvicorn==0.17.6
    # via -r requirements.in
